Remove LPC matching
After reversing AppleLPC, it appears that the offsets do not match our PCH controller
There are too many differences which makes the driver useless. In theory we can patch the driver with good register offsets, but functionality of AppleLPC is not useful for us. It enables ASPM (power savings) as well as restart after power loss (not that useful)