Native TB3 firmware support
With a modded firmware, the controller should behave exactly the same as
a real Mac. As such, we modify the SSDT tables to be the same as on OSX,
including ICM restart on Boot Camp. However, because we broke the firmware
signature, the controller will only operate in safe mode on Windows.

With this change, we see a regression in power savings. This is because we
do not have control over the separate USB/CIO force power GPIO as used on
real Macs. The "force power" GPIO is no longer used.