Bring back legacy Thunderbolt support
The new "native" Thunderbolt patches are better but they require risky NVM
patching as well as break Windows Thunderbolt support. We give the option to use
the older "legacy" patches which uses ICM link management and does not give OSX
control of the LC. The downside is that native Apple TB devices will not work as
well and also we regress to the older stability issues such as kernel panic on
sleep resume when no device is plugged in.