Implement Github Actions (#117)
We use Github's new CI feature to automatically build packages for every commit (that touches the configuration, packages, etc) and pull request. This ensures that packages are built in a consistent environment.

This also makes certain tasks such as updating a kext much easier. For example, if a new version of Lilu comes out, just send a pull request updating source.txt with the latest URL and SHA. Then a package will automatically be built and people can test it. If everything works, a new release is created and the CI system will automatically upload the built asset to the release.