installer: fixed permission issues in Catalina
In Catalina, file system access is granted per-location. Removable devices
are one such location that requires explicit permission to access.

Due to a bug (feature?), even if you grant Installer.app permission to
access removable devices, that permission does not extend to the Shell
scripts the Installer runs as part of the install.

However, due to another bug (feature?) in the permission system, the access
to external drives is limited ONLY to the root directory of the drive. This
means we cannot do a listing on the root directory because we lack the
permission, but we CAN read/write anything beyond that including all files
and subdirectories of the drive. Now some may say this makes the security
feature completely useless because as long as you can guess/brute force the
names of the files/directories on the drive, then you can access it without
any extra permissions. But do not tell Apple this or we would need another
workaround.

So instead of finding InstallInfo.plist automatically, we now hard code a
list of all possible paths to InstallInfo.plist (currently this only
includes "Install macOS Mojave.app" and "Install macOS Catalina.app" but now
we have the unfortunate task of adding to this list with every new OSX
release). Side note: the modding of InstallInfo.plist is itself another
bug/feature because technically the OSX installer is signed and all pkgs
that is included must be signed. But somehow, if you include your pkg under
"Additional Installers", it does not have to be signed. Technically this is
not a security issue because you would need admin priviledges to even change
the OSX installer app, but then why bother blocking unsigned packages in the
other sections? Anyways if Apple "fixes" this then we no longer get an easy
installer patch so our time here is limited anyways and maybe we do not have
to hard code every future OSX installer path anyways.

tl;dr: this fixes #96 and related issues. You no longer need to turn off SIP.