config: improve TB compatibility
Add linkDetails to identify dual-lane support (fixes #149)
Add ThunderboltDROM to fix XDomain support (fixes #147)

AppleThunderboltIP was not working due to a bug where if the NHI "port"
was marked "disabled" by the DROM, then a field returns 0 and underflows.
The fix is to replace the DROM with a patched version that does not have
the NHI port (port 5) disabled. The CRC32 is also updated.

Alternatively, we could use TBPatcher to permanently change DROM but it's
too complex to manage the case of updating patches (and supporting patch
uninstallation).