config: fixed boot issue with multiple EC
On some machines an extra embedded controller device (PNP0C09) named ECDV
is present and causes boot to fail on 10.15. This patch disables both H_EC
and ECDV by replacing PNP0C09 with PNPFFFF. Because a device named "EC"
must exist for AppleBusPowerController matching, we create a fake device
named EC in addition to the other two.

The NUC does not have a real embedded controller so matching the EC driver
to H_EC or ECDV was bad anyways. Our new fake EC will not load the driver.

Fixes #169