installer: skip Apple EFI updates
BlacklistAppleUpdate is not good enough because the Updater will still bless the EFI
to start the updater. Then once it restarts OC will boot into the first non-blessed
entry due to the blacklist. If that is not the Installer, then it breaks the user
experience. We use this EFI variable 'run-efi-updater' to skip EFI updates.